Finish incorporating configuration from separate project, metrics from Indy
Configuration BeanSectionListener refactored to eliminate need for xbean-reflect, and to allow
a setter with a Map parameter for unmatched properties. This is useful when your configuration is
a combination of structured and unstructured parameters.

Metrics has been completely overhauled, to have separate configuration sections
for each reporter type, and to allow mixing and matching via CDI of different
reporters, based on what dependencies are used in the project.

Also, refactored the camel modules into a hierarchy to preserve a common project structure.