Refactor and fold in (#17)
* Merge the old configuration project into propulsor.

* Refactoring project structure and including two external codebases

Refactoring to group similar modules under a module sub-structure for camel.

Incorporating the old configuration project codebase as part of propulsor.

Folding in the Indy metrics subsystem into the metrics module set.

All of this still needs work. This is just a save point while I work.

* Finish incorporating configuration from separate project, metrics from Indy

Configuration BeanSectionListener refactored to eliminate need for xbean-reflect, and to allow
a setter with a Map parameter for unmatched properties. This is useful when your configuration is
a combination of structured and unstructured parameters.

Metrics has been completely overhauled, to have separate configuration sections
for each reporter type, and to allow mixing and matching via CDI of different
reporters, based on what dependencies are used in the project.

Also, refactored the camel modules into a hierarchy to preserve a common project structure.