Fixing NPEs, sanitizing interpolation, fixing coercion
We had a reference to the YAML config format, which I haven't finished, so I took it out.

There were a couple of NPE places after I replaced the old xbeans-reflect code, so I fixed those.

Interpolation was happening BOTH when we read the config file with its logic to inline
included files AND in the main configuration process. I removed the interpolation in the
config-file inlining process to reduce opportunity for nasty surprises.

I added a method in ConfigUtils called loadStandardConfigProperties() which loads
System.getProperties and then overlays System.getenv(), with all envar keys prefixed
by 'env.'.

Finally, we had some problems where setters accepted primitive types, and coercion was failing.
I fixed using Integer.TYPE and so on.